back/requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-oauth-toolkit==1.3.3
# django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
whitenoise==5.2.0
daphne==3.0.1
psycopg2-binary==2.8.6
httpx==0.16.1
django-environ==0.4.5
coreapi==2.3.3
# djoser==2.1.0
rest-social-auth==5.0.1
djoser
